+++ src/app/core/type/language.type.ts
@@ -1,0 +1,13 @@
+export enum UiLanguage {
+    English = 1,
+    Portuguese = 2,
+    Russian = 3,
+    Thai = 4,
+    German = 5,
+    French = 6,
+    Spanish = 7,
+    Korean = 8,
+    SimplifiedChinese = 9,
+    TraditionalChinese = 10,
+    Polish = 11,
+}