+++ src/app/data/github/schema/releases.type.ts
@@ -1,9 +1,0 @@
-export interface Release {
-    tag_name: string;
-    html_url: string;
-    assets: ReleaseAsset[];
-}
-
-export interface ReleaseAsset {
-    browser_download_url: string;
-}