+++ src/app/data/poe/service/trade-http.service.ts
@@ -91,6 +91,12 @@ export class TradeHttpService {
             case Language.Korean:
                 baseUrl = environment.poe.koreanUrl;
                 break;
+            case Language.SimplifiedChinese:
+                baseUrl = environment.poe.simplifiedChineseUrl;
+                break;
+            case Language.TraditionalChinese:
+                baseUrl = environment.poe.traditionalChineseUrl;
+                break;
             default:
                 throw new Error(`Could not map baseUrl to language: '${language}'.`);
         }