+++ src/app/layout/page/overlay/overlay.component.ts
@@ -1,13 +1,13 @@
 import { ChangeDetectionStrategy, Component, HostListener, Inject, OnDestroy, OnInit } from '@angular/core';
-import { AppService, BrowserService, RendererService, WindowService } from '@app/service';
+import { AppService, RendererService, WindowService } from '@app/service';
 import { DialogRefService } from '@app/service/dialog';
 import { ShortcutService } from '@app/service/input';
 import { FEATURE_MODULES } from '@app/token';
-import { AppUpdateState, FeatureModule, VisibleFlag } from '@app/type';
+import { AppUpdateState, FeatureModule, UiLanguage, VisibleFlag } from '@app/type';
 import { TranslateService } from '@ngx-translate/core';
 import { SnackBarService } from '@shared/module/material/service';
 import { ContextService } from '@shared/module/poe/service';
-import { Context, Language } from '@shared/module/poe/type';
+import { Context } from '@shared/module/poe/type';
 import { BehaviorSubject, Observable } from 'rxjs';
 import { distinctUntilChanged, flatMap, tap } from 'rxjs/operators';
 import { UserSettingsService } from '../../service/user-settings.service';
@@ -37,7 +37,7 @@ export class OverlayComponent implements OnInit, OnDestroy {
     private readonly shortcut: ShortcutService,
     private readonly dialogRef: DialogRefService,
     private readonly translate: TranslateService) {
-    this.translate.setDefaultLang(`${Language.English}`);
+    this.translate.setDefaultLang(`${UiLanguage.English}`);
   }
 
   @HostListener('window:beforeunload', [])
@@ -56,7 +56,7 @@ export class OverlayComponent implements OnInit, OnDestroy {
 
   private initSettings(): void {
     this.userSettingsService.init(this.modules).subscribe(settings => {
-      this.translate.use(`${settings.language}`);
+      this.translate.use(`${settings.uiLanguage}`);
       this.displayVersion$.next(settings.displayVersion);
       this.window.setZoom(settings.zoom / 100);
       this.context.init(this.getContext(settings));
@@ -117,7 +117,7 @@ export class OverlayComponent implements OnInit, OnDestroy {
       ).subscribe(settings => {
         this.userSettingsOpen = null;
 
-        this.translate.use(`${settings.language}`);
+        this.translate.use(`${settings.uiLanguage}`);
         this.window.setZoom(settings.zoom / 100);
         this.displayVersion$.next(settings.displayVersion);
         this.context.update(this.getContext(settings));