+++ src/app/shared/module/poe/service/item/item-external.service.ts
@@ -40,6 +40,8 @@ const LANGUAGE_MAPPING = {
     [Language.Russian]: 'ru',
     [Language.Spanish]: 'sp',
     [Language.Thai]: 'th',
+    [Language.SimplifiedChinese]: 'cn',
+    [Language.TraditionalChinese]: 'tw',
 }
 
 @Injectable({